// src/types/Player.ts
export interface Player {
  "1/3": number;
  "2CrdY": number;
  "90s": number;
  "#OPA": number | null;
  "#OPA/90": number | null;
  "+/-": number;
  "+/-90": number;
  "/90": number | null;
  "A-xAG": number;
  Age: number;
  Ast: number;
  Att: number;
  "Att 3rd": number;
  "Att (GK)": number | null;
  "Att Pen": number;
  AvgDist: number | null;
  AvgLen: number | null;
  Blocks: number;
  Born: number;
  CK: number;
  CPA: number;
  CS: number | null;
  "CS%": number | null;
  Carries: number;
  Clr: number;
  Cmp: number;
  "Cmp%": number;
  Comp: string;
  Compl: number;
  CrdR: number;
  CrdY: number;
  Crs: number;
  CrsPA: number;
  D: number | null;
  Dead: number;
  Def: number;
  "Def 3rd": number;
  "Def Pen": number;
  Dis: number;
  Dist: number | null;
  Err: number;
  FK: number;
  Fld: number;
  Fls: number;
  "G+A": number;
  "G+A-PK": number;
  "G-PK": number;
  "G-xG": number;
  "G/Sh": number | null;
  "G/SoT": number | null;
  GA: number | null;
  GA90: number | null;
  GCA: number;
  GCA90: number;
  Gls: number;
  In: number;
  Int: number;
  KP: number;
  L: number | null;
  "Launch%": number | null;
  Live: number;
  Lost: number;
  MP: number;
  "Mid 3rd": number;
  Min: number;
  "Min%": number;
  Mis: number;
  "Mn/MP": number;
  "Mn/Start": number;
  "Mn/Sub": number;
  Nation: string;
  OG: number;
  Off: number;
  "On-Off": number;
  Opp: number | null;
  Out: number;
  PK: number;
  PKA: number | null;
  PKatt: number;
  PKcon: number;
  PKm: number | null;
  PKsv: number | null;
  PKwon: number;
  PPA: number;
  PPM: number;
  PSxG: number | null;
  "PSxG+/-": number | null;
  "PSxG/SoT": number | null;
  Pass: number;
  PassDead: number;
  PassLive: number;
  Player: string;
  Pos: string;
  PrgC: number;
  PrgDist: number;
  PrgP: number;
  PrgR: number;
  Rec: number;
  Recov: number;
  Rk: number;
  SCA: number;
  SCA90: number;
  "Save%": number | null;
  Saves: number | null;
  Sh: number;
  "Sh/90": number;
  SoT: number;
  "SoT%": number | null;
  "SoT/90": number;
  SoTA: number | null;
  Squad: string;
  Starts: number;
  Stp: number | null;
  "Stp%": number | null;
  Str: number;
  Subs: number;
  Succ: number;
  "Succ%": number;
  Sw: number;
  TB: number;
  TI: number;
  TO: number;
  Thr: number | null;
  Tkl: number;
  "Tkl%": number;
  "Tkl+Int": number;
  TklW: number;
  Tkld: number;
  "Tkld%": number;
  TotDist: number;
  Touches: number;
  W: number | null;
  Won: number;
  "Won%": number | null;
  "np:G-xG": number;
  npxG: number;
  "npxG+xAG": number;
  "npxG/Sh": number | null;
  onG: number;
  onGA: number;
  onxG: number;
  onxGA: number;
  unSub: number;
  xA: number;
  xAG: number;
  xG: number;
  "xG+/-": number;
  "xG+/-90": number;
  "xG+xAG": number;
}